Add files via upload
Details are within the code. I hope to tweak the code so it's easier to configure for each cell. Currently, you'll have to change the code before you upload it to match the cell. (I.e. correct the code for cell 1, upload to ESP-12E #1. Correct the code for Cell 2 then upload to ESP-12E #2.)  Note: This code only manages battery voltage uploading (to adafrutit io) and battery balancing. One of the ESP-12E's should have code added to control the overall load (or charger) itself. When any cell goes below (or above) a set value you need to turn off the load (or the charger).